+++ docs/whats-new.md
@@ -1,3 +1,9 @@
+## UNRELEASED
+
+### Feat
+- added spatial/temporal subsetting and retrieving of insitu observation data with `dfmt.ssh_catalog_subset()` and `dfmt.ssh_retrieve_data()` by [@veenstrajelmer](https://github.com/veenstrajelmer) in [#711](https://github.com/Deltares/dfm_tools/pull/711)
+
+
 ## 0.18.0 (2023-12-08)
 
 ### Feat